Improve tests organization
